Organised the components
Created separate folders in the components folder and moved the corresponding component files into the folders.
