+++ client/src/components/Duel/DuelLoad.tsx
