+++ client/src/components/Other/DisplayLogo.tsx
