+++ client/src/components/Other/EndTournament.tsx
