+++ client/src/components/Other/WaitingToStart.tsx
