+++ client/src/components/Player/StarsInfo.tsx
