+++ client/src/components/Spectator/SpectatorCount.tsx
