+++ client/src/pages/JoinGame.tsx
@@ -1,4 +1,4 @@
-import DisplayLogo from "../components/DisplayLogo.tsx";
+import DisplayLogo from "../components/Other/DisplayLogo.tsx";
 import { useEffect, useState } from "react";
 
 import { socket } from "../App.tsx";
